requirements-beauty.txt
# Lambda deployment requirements for Beauty Lambda
# Personal Color + Hair Color (NO PyTorch)

# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
boto3==1.35.0

# Redis Caching
redis==5.0.1

# Google Gemini AI
google-generativeai==0.3.2
google-genai==1.0.0

# Image Processing (NO PyTorch, NO scikit-learn for ML)
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
mediapipe==0.10.9

# Lambda Handler
mangum==0.17.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
pybreaker==1.0.1
slowapi==0.1.9
psutil==5.9.6

# Monitoring
sentry-sdk[fastapi]==1.40.0
